refactor: move send  verification token to email service use case
