+++ src/use-cases/errors/index.ts
@@ -1,2 +1,1 @@
 export * from '@/use-cases/errors/unauthorized-error';
-export * from '@/use-cases/errors/email-service-error';
