+++ src/use-cases/services/email-service/errors/email-service-error.ts
