+++ src/use-cases/services/email-service/errors/index.ts
@@ -1,0 +1,1 @@
+export * from '@/use-cases/services/email-service/errors/email-service-error';
