+++ src/use-cases/services/email-service/interfaces/email-options-interface.ts
