+++ src/use-cases/services/email-service/interfaces/email-service-interface.ts
