+++ src/use-cases/services/email-service/interfaces/index.ts
@@ -1,0 +1,2 @@
+export * from '@/use-cases/services/email-service/interfaces/email-options-interface';
+export * from '@/use-cases/services/email-service/interfaces/email-service-interface';
