+++ src/use-cases/services/email-service/send-email/index.ts
@@ -1,0 +1,1 @@
+export * from '@/use-cases/services/email-service/send-email/send-email-service';
