--- src/use-cases/verification-token/send-verification-token/index.ts
@@ -1,1 +1,0 @@
-export * from '@/use-cases/verification-token/send-verification-token/send-verification-token-use-case';
